requirements.txt
openai
requests
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
typer
ipykernel
python-dotenv
